requirements.txt
openai>=1.68.0
opentelemetry-api>=1.32.0
opentelemetry-exporter-otlp>=1.32.0
opentelemetry-sdk>=1.32.0
PyYAML>=6.0.2
pytest>=8.3.5
